+++ packages/good/test/multi_scene_test.dart
@@ -95,7 +95,7 @@ class _MultiGame extends Game {
   @override
   void describeScenes(GameSceneDescriptor descriptor) {
     super.describeScenes(descriptor);
-    level = descriptor.has(_Level());
+    level = descriptor.has(_Level.new);
   }
 }